Cambiando cantidad a Double

+++ src/main/java/com/example/buensaboruno/domain/dtos/ArticuloManufacturadoDetalleDTO.java
@@ -14,6 +14,6 @@ import lombok.Setter;
 @Setter
 public class ArticuloManufacturadoDetalleDTO extends BaseDTO {
 
-    private Integer cantidad;
+    private Double cantidad;
     private ArticuloInsumoDTO articuloInsumo;
 }
